Move: 관리자의 포토카드 목록 페이지에서 사용되는 컴포넌트들을 @component/photo-list 로 이동
