+++ client/src/app/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from "@reduxjs/toolkit";
 import sidebarSlice from "@component/sidebar/sidebarSlice";
-import photoSearchSlice from "@page/admin/photo/list/searchSlice";
+import photoSearchSlice from "@component/photo-list/content/searchSlice";
 import authSlice from "@util/auth/authSlice";
 
 export default combineReducers({
