+++ client/src/component/photo-list/basic/SkeletonPhotoCard.tsx
