+++ client/src/component/photo-list/content/SearchInput.tsx
