+++ client/src/component/sidebar/content/UserContent.tsx
@@ -17,9 +17,9 @@ function UserContent({ children }: UserContentProps & typeof UserContentDefaultP
       <ul className="category">
         <li className="subheader-label">포토카드 교환</li>
         <ParentItem id="trade" icon={faShareNodes} text="교환" iconMarginRight="0.95em">
-          <ChildItem to="/trade/writer" text="등록" />
-          <ChildItem to="/trade/search" text="찾기" />
-          <ChildItem to="/trade/matching" text="매칭" />
+          <ChildItem to="/trade/search" text="포토카드 찾기" />
+          <ChildItem to="/trade/writer" text="교환글 등록" />
+          <ChildItem to="/trade/matching" text="교환 매칭" />
         </ParentItem>
         <ChildItem to="/trade/history" icon={faTimeline} text="내역" iconMarginRight="0.6em"/>
       </ul>
