+++ client/src/page/admin/photo/list/Index.tsx
@@ -2,13 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
 import Button from '@component/form/Button';
-import Card from '@component/card/basic/Card';
-import CardHeader from '@component/card/basic/CardHeader';
-import CardBody from '@component/card/basic/CardBody';
-import SearchInput from './SearchInput';
-import SearchLabelList from './SearchLabelList';
-import PhotoList from './PhotoList';
-import FilterCheck from './FilterCheck';
+import PhotoListContainer from '@component/photo-list/PhotoListContainer';
 import './Index.scss';
 
 interface PhotoListPageProps {
@@ -32,16 +26,7 @@ function PhotoListPage({ children }: PhotoListPageProps & typeof PhotoListPageDe
           >추가</Button>
         </Link>
       </section>
-      <Card>
-        <CardHeader padding="0" borderBottom="0">
-          <SearchInput />
-          <SearchLabelList />
-        </CardHeader>
-        <CardBody>
-          <FilterCheck />
-          <PhotoList />
-        </CardBody>
-      </Card>
+      <PhotoListContainer />
     </section>
   );
 }
